+++ x/bank/simulation/operations.go
@@ -174,7 +174,7 @@ func sendMsgSend(
 		return err
 	}
 
-	_, _, err = app.SimTxFinalizeBlock(txGen.TxEncoder(), tx)
+	_, _, err = app.SimDeliver(txGen.TxEncoder(), tx)
 	if err != nil {
 		return err
 	}
@@ -390,7 +390,7 @@ func sendMsgMultiSend(
 	if err != nil {
 		return err
 	}
-	_, _, err = app.SimTxFinalizeBlock(txGen.TxEncoder(), tx)
+	_, _, err = app.SimDeliver(txGen.TxEncoder(), tx)
 	if err != nil {
 		return err
 	}